Added math for new summary fields and updated tests

+++ src/Lykke.AlgoStore.CSharp.AlgoTemplate.Core/Services/IStatisticsService.cs
@@ -9,7 +9,7 @@ namespace Lykke.AlgoStore.CSharp.AlgoTemplate.Core.Services
     {
         void OnQuote(IAlgoQuote quote);
         void OnAction(bool isBuy, double volume, double price);
-        void OnAlgoStarted(decimal initialWalletBalance, decimal assetOneBalance, decimal assetTwoBalance);
+        void OnAlgoStarted(double initialWalletBalance, double assetOneBalance, double assetTwoBalance);
         void OnAlgoStopped();
     }
 }
